⚡️ dnd hook으로 분할
